FALCON-1826 Execution order not honoured when instances are KILLED
The problem was that the unregister for schedule did not unregister the instance immediately. It would add to instancesToIgnore and would get ignored while scheduling. This causes issues as instancesToIgnore will need to be scanned at more than one place.

The fix is to remove the instance from "awaitingSchedule" list synchronously for unregister.

Author: Pallavi Rao <pallavi.rao@inmobi.com>

Reviewers: "Pavan Kumar Kolamuri <pavan.kolamuri@gmail.com>"

Closes #58 from pallavi-rao/1826 and squashes the following commits:

9ace53d [Pallavi Rao] FALCON-1826 Addressed review comments
dec29c0 [Pallavi Rao] FALCON-1826 Execution order not honoured when instances are KILLED