FALCON-1776 Process instance update in titan DB based on JMS notifications on coordinator actions
In JMS notifications from the coordinator action, it is not guaranteed to provide information on timestamp and user workflow version. Need to allow these properties to be optional when adding process instance to titan DB. Also need to address the different format when there is no input/output feed. See the picture below for an example result (i.e. WAITING instance).

<img width="1083" alt="screen shot 2016-08-12 at 4 50 59 pm" src="https://cloud.githubusercontent.com/assets/10202347/17640302/aec16864-60b1-11e6-9a38-b6dc5cd8b890.png">

Author: yzheng-hortonworks <yzheng@hortonworks.com>

Reviewers: "sandeep <sandysmdl@gmail.com>, Peeyush<peeyushb@apache.org>, Balu <balu@apache.org>"

Closes #263 from yzheng-hortonworks/FALCON-1776