FALCON-2133 Disable the LogMover from FeedReplication on secure mode.
LogMover is only supported in case of un-secure mode. It has to be disable in the secure mode. In FeedReplication test cases, LogMover assertions will be done after the functionality tested covered. In this cases, we need to verify the functionality but only disable LogMover.

Author: Murali Ramasami <mramasami@hortonworks.com>

Reviewers: "Pragya Mittal <mittal.pragya23@gmail.com>"

Closes #277 from muraliramasami/master and squashes the following commits:

8e5a37a [Murali Ramasami] FALCON-2133: Disable the LogMover assertions from FeedReplication on Secure Mode
c1a8b60 [Murali Ramasami] FALCON-2133: Disable the LogMover assertions from FeedReplication on Secure Mode